test+fix: canvas-stability + scroll-virt specs uncover real bugs
Tier-1 of the test plan. New Playwright specs that drive the live
playground via chrome-devtools-style instrumentation, plus the bugs
they immediately exposed.

NEW SPECS
- canvas-stability.spec.ts: MutationObserver on canvas attrs +
  getImageData center-pixel sampling. Assertion: canvas alpha is never
  observably 0 (cleared) during window resize, column drag-resize,
  or scroll. Catches the head-to-toe flicker class.
- scroll-virtualization.spec.ts: drives the 10M-row preset and asserts
  (1) virtualization engages (scrollScale > 1, totalHeight > 16 Mpx),
  (2) physical scrollbar at max reaches lastVisibleRow = numRows - 1,
  (3) scrollToRow lands the target inside the viewport for {0, 1k,
  100k, 1M, 5M, 9.99M}, (4) visible row index never regresses during
  forward scroll.
- canvas-stability.ts: reusable assertCanvasStable() helper.

BUGS FOUND + FIXED
1. FenwickHeights used Float32Array, which loses integer precision past
   2^24 ≈ 16.7M. For 10M rows × ~28 px = 280 Mpx, the tree's partial
   sums (stored Float32) accumulate rounding error and totalHeight
   drifted to 28.8M instead of 288M — visible-row meter stalled at
   row 999,999 even at the physical bottom. Switched tree + heights
   to Float64Array (53-bit mantissa, exact for any realistic dataset).
   Cost: 40 → 80 MB at 10M rows. Worth it.

2. Grid.setRowSource updated this.fenwick but not this.baseHeights.
   When React subsequently fired setExpanded (on mode / numRows
   change because expandedRows state resets), rebuildHeightsFromExpansion
   reconstructed fenwick from the STALE baseHeights — instantly
   reverting fenwick size to the prior dataset's row count. Symptom:
   meter capped at the previous dataset's last row even after switching
   to a larger one. Fix: setRowSource now also assigns baseHeights
   from the new rowHeight argument.

NEW API: Grid.getViewportInfo() — public method returning logical
scrollTop / scrollScale / numRows / totalHeight / firstVisibleRow /
lastVisibleRow / viewport dims. Used by the scroll-virt specs to
assert on internal state. Exposed via window.__onegrid in playground.

METHODOLOGY
These bugs were caught by writing actual specs that drove the live
browser. Static reading missed both of them across multiple sessions.
The test-coverage gap was real — these specs close it for the scroll
and resize hot paths. Tier-2 (visual regression, cross-browser matrix,
mode-matrix coverage) is the next step.

+++ packages/data/src/fenwick.ts
@@ -8,8 +8,19 @@
 //   indexAtOffset(y)    — first row whose top edge is > y    — O(log n)
 //   setHeight(i, h)     — replace row i's height             — O(log n)
 //
-// Storage is a Float32Array of size n + 1 (1-indexed). 10M rows uses ~40 MB,
+// Storage is a Float64Array of size n + 1 (1-indexed). 10M rows uses ~80 MB,
 // significant but tolerable, with zero allocations on the scroll hot path.
+//
+// Why Float64, not Float32: Float32 has 24 mantissa bits, exact for integers
+// up to 2^24 ≈ 16.7M. At 10M rows × ~28 px avg = 280 Mpx total height, the
+// partial sums in the Fenwick tree exceed Float32's exact-integer range and
+// each `tree[i] = tree[i] + delta` operation rounds at ~16 ULP. After
+// millions of adds the accumulated drift makes totalHeight wrong by orders
+// of magnitude — observed empirically as totalHeight reporting 28.8M
+// instead of 288M for a 10M-row dataset, with the visible-row meter
+// stalling at row 999,999 even when scrolled to the physical bottom.
+// Float64 has 53 mantissa bits, exact for integers up to 2^53 ≈ 9e15 px,
+// which covers any realistic row count × row height.
 //
 // CodeMirror 6 uses the same data structure for line heights; the technique
 // scales to multi-million-row spreadsheets without measurable cost on each
@@ -17,14 +28,14 @@
 // =============================================================================
 
 export class FenwickHeights {
-  private readonly tree: Float32Array;
-  private readonly heights: Float32Array;
+  private readonly tree: Float64Array;
+  private readonly heights: Float64Array;
   public readonly length: number;
 
   constructor(heights: ArrayLike<number>) {
     this.length = heights.length;
-    this.heights = new Float32Array(heights.length);
-    this.tree = new Float32Array(heights.length + 1);
+    this.heights = new Float64Array(heights.length);
+    this.tree = new Float64Array(heights.length + 1);
     for (let i = 0; i < heights.length; i++) {
       const h = heights[i] ?? 0;
       this.heights[i] = h;
